feat: Save match in DB
